Align DisplayParentUpdatesMojoTest and UpdateParentMojoTest
