--- versions-maven-plugin/src/test/resources/org/codehaus/mojo/display-parent-updates/no-parent/pom.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>display-parent-updates</groupId>
-  <artifactId>project-has-no-parent</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-</project>
