--- versions-maven-plugin/src/test/resources/org/codehaus/mojo/display-parent-updates/parent/pom.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>default-group</groupId>
-    <artifactId>artifactA</artifactId>
-    <version>1.0.0</version>
-  </parent>
-
-  <artifactId>display-parent-updates-test</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-</project>
